myapp/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
djangorestframework==3.13.1
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
nose==1.3.7
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.2
PyYAML==6.0
Random-Word==1.0.7
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.11
psycopg2==2.9.3
celery==5.2.7
django-silk==5.0.1
django-redis==5.2.0
elasticsearch-dsl==7.4.0
django-elasticsearch-dsl==7.2.2